requirements.txt
kafka-python>=2.0.2
redis>=5.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
prometheus-client>=0.19.0
structlog>=23.2.0
pyyaml>=6.0.1
